pyproject.toml
[project]
name = "stt-llm-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastrtc[stt]>=0.0.19",
    "kokoro-onnx>=0.4.7",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
]
